todo12

+++ Frontend/src/Components/todo/Todo.js
@@ -4,10 +4,8 @@ import TodoCards from "./TodoCards";
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Update from './Update';
-// import {useSelector} from "react-redux";
-// import {useDispatch} from 'react-redux'
-// import{authActions} from "../../store/index";
 import axios from "axios";
+let toUpdateArray = [];
 let id = sessionStorage.getItem("id");
 const Todo = () => {
   const [Inputs, setInputs] = useState({title:"",body:""});
@@ -15,9 +13,12 @@ const Todo = () => {
   
   
 const del = async(Cardid) =>{
+ if(id){
   await axios.delete(`http://localhost:1000/api/v2/deleteTask/${Cardid}`,{data:{id:id},}).then(()=>{
     toast.success("Your task is deleted");
   });
+ }
+ 
  
 }
 const show=()=>{
@@ -52,15 +53,23 @@ const show=()=>{
   console.log(value);
   document.getElementById("todo-update").style.display = value;
  }
- console.log(Array);
+
  useEffect(() => {
-  const fetch = async() => {
-    await axios.get(`http://localhost:1000/api/v2/getTasks/${id}`).then((response)=>{
-      setArray(response.data.list);
-    })
+  if (id) {
+    const fetch = async () => {
+      await axios.get(`http://localhost:1000/api/v2/getTasks/${id}`).then((response) => {
+        setArray(response.data.list);
+        // console.log(Array);
+      });
+    };
+    fetch();
   }
-  fetch();
 }, [submit]);
+const update =(value)=>{
+  toUpdateArray=Array[value];
+}
+
+
   return (
     <>
     <div className="todo">
@@ -86,7 +95,7 @@ const show=()=>{
             <div className="row">
                
             {Array && Array.map((item, index) => (<div className='col-lg-3 col-8 mx-5 my-2' key={index}>
-              <TodoCards title={item.title} body={item.body} id={item._id} delid={del} display={dis}/>
+              <TodoCards title={item.title} body={item.body} id={item._id} delid={del} display={dis} updateId={index} toBeUpdate={update}/>
               </div>)
             )}
             
@@ -100,7 +109,7 @@ const show=()=>{
     </div>
     <div className="todo-update bg-success " id="todo-update">
       <div className="container">
-        <Update display={dis}/>
+        <Update display={dis} update={toUpdateArray}/>
         </div>
         </div>
     </>
